Education tools and wizard permission DONE

+++ routes/web.php
@@ -339,22 +339,22 @@ Route::middleware(['auth', 'roles:admin', 'check.blocked.ip'])->group(function (
 
         Route::post('update-subject/{id}', [EducationController::class, 'updateSubject'])->name('update.subject');
 
-        Route::post('delete-grade/{id}', [EducationController::class, 'deleteGrade'])->name('delete.grade');
+        Route::post('delete-grade/{id}', [EducationController::class, 'deleteGrade'])->name('delete.grade')->middleware('permission:education.manageGradeSubject.gradeDelete');
 
-        Route::post('delete-subject/{id}', [EducationController::class, 'deleteSubject'])->name('delete.subject');
+        Route::post('delete-subject/{id}', [EducationController::class, 'deleteSubject'])->name('delete.subject')->middleware('permission:education.manageGradeSubject.subjectDelete');
         
-        Route::get('/add/tools', [EducationController::class, 'AddTools'])->name('add.education.tools');
+        Route::get('/add/tools', [EducationController::class, 'AddTools'])->name('add.education.tools')->middleware('permission:education.manageTools.add');
 
         Route::post('/store/tools', [EducationController::class, 'StoreTools'])->name('store.education.tools');
 
         // Route to show the form for editing a specific tool (edit)
-        Route::get('/tools/{id}/edit', [EducationController::class, 'editTools'])->name('tools.edit');
+        Route::get('/tools/{id}/edit', [EducationController::class, 'editTools'])->name('tools.edit')->middleware('permission:education.manageTools.edit');
 
         // Route to update a specific tool (update)
         Route::put('/tools/{id}', [EducationController::class, 'updateTools'])->name('tools.update');
 
         // Route to delete a specific tool (destroy)
-        Route::delete('/tools/{id}', [EducationController::class, 'destroyTools'])->name('tools.destroy');
+        Route::delete('/tools/{id}', [EducationController::class, 'destroyTools'])->name('tools.destroy')->middleware('permission:education.manageTools.delete');
 
     });
